+++ Cargo.toml
@@ -22,6 +22,7 @@ byteorder = { version = "1.5.0", default-features = false }
 chrono = { version = "0.4", default-features = false }
 clap = { version = "4.5", default-features = false }
 constant_time_eq = { version = "0.3.1" }
+fail = { version = "0.5.1", default-features = false }
 futures = { version = "0.3.31", default-features = false }
 gcp-bigquery-client = { version = "0.25.0", default-features = false }
 # gcp-bigquery-client = { git = "https://github.com/imor/gcp-bigquery-client", default-features = false, rev = "d9fe29a33f9e4dc12c4adf061035ee1628da5e39" }
