+++ etl/Cargo.toml
@@ -13,6 +13,15 @@ required-features = ["bigquery"]
 name = "table_copies"
 harness = false
 
+[[test]]
+name = "integration"
+path = "tests/integration/mod.rs"
+
+[[test]]
+name = "failpoints"
+path = "tests/failpoints/mod.rs"
+required-features = ["fail/failpoints"]
+
 [dependencies]
 config = { workspace = true }
 postgres = { workspace = true, features = ["tokio", "replication"] }
@@ -22,6 +31,7 @@ bigdecimal = { workspace = true, features = ["std"] }
 bytes = { workspace = true }
 byteorder = { workspace = true }
 chrono = { workspace = true, features = ["serde"] }
+fail = { workspace = true, features = ["failpoints"] }
 futures = { workspace = true }
 gcp-bigquery-client = { workspace = true, optional = true, features = [
     "rust-tls",
