+++ etl/tests/failpoints/mod.rs
@@ -1,0 +1,1 @@
+mod pipeline_test;
