+++ etl/tests/mod.rs
@@ -1,2 +1,3 @@
 mod common;
+mod failpoints;
 mod integration;
